+++ backend/package.json
@@ -33,7 +33,7 @@
         "typescript": "^4.1.3"
     },
     "dependencies": {
-        "@material-ui/core": "^4.11.3",
+        "@material-ui/core": "^4.11.4",
         "@material-ui/icons": "^4.11.2",
         "@material-ui/lab": "^4.0.0-alpha.58",
         "@material-ui/system": "^4.11.3",
