+++ backend/resources/ts/place/components/molecules/PlaceCardItemSkeleton.tsx
@@ -1,0 +1,21 @@
+// import React, { FC } from 'react';
+// import CardItem from '@material-ui/core/CardItem';
+// import CardItemIcon from '@material-ui/core/CardItemIcon';
+// import CardItemText from '@material-ui/core/CardItemText';
+// import Skeleton from '@material-ui/lab/Skeleton';
+// import AssignmentIcon from '@material-ui/icons/Assignment';
+
+// const PlaceCardItemSkeleton: FC = () => (
+//   <CardItem>
+//     <CardItemIcon>
+//       <Skeleton>
+//         <AssignmentIcon />
+//       </Skeleton>
+//     </CardItemIcon>
+//     <Skeleton width="100%">
+//       <CardItemText primary="skeleton" secondary="skeleton" />
+//     </Skeleton>
+//   </CardItem>
+// );
+
+// export default PlaceCardItemSkeleton;
