+++ backend/resources/ts/place/hooks/index.tsx
@@ -1,4 +1,4 @@
-export { default as useGetPlaceListQuery } from './useGetPlaceListQuery';
+export { default as useGetPlaceCardQuery } from './useGetPlaceCardQuery';
 export { default as useGetPlaceQuery } from './useGetPlaceQuery';
 export { default as usePostPlaceMutation } from './usePostPlaceMutation';
 // export { default as usePatchPlaceMutation } from './usePatchPlaceMutation';
