+++ backend/resources/ts/place/types/Place.ts
@@ -1,10 +1,16 @@
+
+import { User } from '../../user/types/User';
+
 export type Place = {
     data: { key: string[] }
     id: number
     name: string
     address: string
     comment: string
-    created_at: Date
-    updated_at: Date
+    createdAt: Date
+    updatedAt: Date
+    placeImages: string
     placeId: string
+    user: User
+    tags: string
 }
