+++ backend/resources/ts/user/types/User.ts
@@ -1,4 +1,13 @@
 export type User = {
-  name: string;
-  authType: 'Social' | 'Mail';
-};
+  id: number
+  name: string
+  authType: 'Social' | 'Mail'
+  email: string
+  emailVerifiedAt: boolean
+  age: number
+  userImage: HTMLImageElement
+  introduction: string
+  createdAt: Date
+  updatedAt: Date
+  deletedAt: Date
+}
