+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-bootstrap": "^1.6.1"
+  }
+}
